Update Memory_screen.m
I did the loops for the feedback with Andra's keycodes ("d" for old response and "k" for new response). I think because we don't have any images yet, it doesn't do anything on the screen when I run the code. If it still doesn't work once we insert the images, I'll look into it :) But for now I just wanted to put it in there just in case!